+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/commands/admin/CleanUpChatCommand.java
@@ -1,7 +1,7 @@
 package de.relluem94.minecraft.server.spigot.essentials.commands.admin;
 
 import de.relluem94.minecraft.server.spigot.essentials.commands.Admin;
-import de.relluem94.minecraft.server.spigot.essentials.constants.MessageKey;
+import de.relluem94.minecraft.server.spigot.essentials.enums.MessageKey;
 import de.relluem94.minecraft.server.spigot.essentials.interfaces.SubCommand;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
