+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/commands/modify/ClipboardCommand.java
@@ -2,10 +2,10 @@ package de.relluem94.minecraft.server.spigot.essentials.commands.modify;
 
 import de.relluem94.minecraft.server.spigot.essentials.RelluEssentials;
 import de.relluem94.minecraft.server.spigot.essentials.commands.Modify;
-import de.relluem94.minecraft.server.spigot.essentials.constants.MessageKey;
-import de.relluem94.minecraft.server.spigot.essentials.helpers.objects.Selection;
-import de.relluem94.minecraft.server.spigot.essentials.helpers.pojo.ModifyClipboardEntry;
+import de.relluem94.minecraft.server.spigot.essentials.enums.MessageKey;
 import de.relluem94.minecraft.server.spigot.essentials.interfaces.SubCommand;
+import de.relluem94.minecraft.server.spigot.essentials.model.Selection;
+import de.relluem94.minecraft.server.spigot.essentials.model.pojo.ModifyClipboardEntry;
 import de.relluem94.rellulib.stores.DoubleStore;
 import org.bukkit.entity.Player;
 import org.jspecify.annotations.NonNull;
