+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/events/bag/BlockDropItemBags.java
@@ -5,7 +5,7 @@ import de.relluem94.minecraft.server.spigot.essentials.RelluEssentials;
 import de.relluem94.minecraft.server.spigot.essentials.helpers.BagHelper;
 import de.relluem94.minecraft.server.spigot.essentials.helpers.EnchantmentHelper;
 import de.relluem94.minecraft.server.spigot.essentials.helpers.ItemHelper;
-import de.relluem94.minecraft.server.spigot.essentials.helpers.pojo.PlayerEntry;
+import de.relluem94.minecraft.server.spigot.essentials.model.pojo.PlayerEntry;
 import de.relluem94.rellulib.stores.DoubleStore;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
